Move Utxostore rules to Utxoset folder
