+++ src/Blockcore.Features.Consensus/Rules/CommonRules/TransactionDuplicationActivationRule.cs
@@ -2,6 +2,7 @@
 using Blockcore.Base.Deployments;
 using Blockcore.Consensus;
 using Blockcore.Consensus.Rules;
+using Blockcore.Features.Consensus.Rules.UtxosetRules;
 using Blockcore.Utilities;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
